web/ui/pyproject.toml
[tool.poetry]
name = "ui"
version = "0.1.0"
description = ""
authors = ["juyony <wnsdyd1991@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.33.0"
requests = "^2.31.0"
pandas = "^2.2.2"
matplotlib = "^3.8.4"
pillow = "^10.3.0"
st-pages = "^0.4.5"
streamlit-authenticator = "^0.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
